+++ lizmap/forms/edition_edition.py
@@ -1,6 +1,11 @@
 """Dialog for edition layer edition."""
 
-from qgis.core import QgsMapLayerProxyModel, QgsProject, QgsWkbTypes
+from qgis.core import (
+    QgsMapLayerProxyModel,
+    QgsProject,
+    QgsProviderRegistry,
+    QgsWkbTypes,
+)
 from qgis.PyQt.QtWidgets import QMessageBox
 
 from lizmap.definitions.definitions import LwcVersions
@@ -8,7 +13,6 @@ from lizmap.definitions.edition import EditionDefinitions, layer_provider
 from lizmap.forms.base_edition_dialog import BaseEditionDialog
 from lizmap.qgis_plugin_tools.tools.i18n import tr
 from lizmap.qgis_plugin_tools.tools.resources import load_ui
-from lizmap.tools import excluded_providers
 
 __copyright__ = 'Copyright 2020, 3Liz'
 __license__ = 'GPL version 3'
@@ -52,7 +56,9 @@ class EditionLayerDialog(BaseEditionDialog, CLASS):
 
         self.layer.layerChanged.connect(self.layer_changed)
         self.layer.setFilters(QgsMapLayerProxyModel.VectorLayer)
-        self.layer.setExcludedProviders(excluded_providers())
+        providers = QgsProviderRegistry.instance().providerList()
+        providers.remove('postgres')
+        self.layer.setExcludedProviders(providers)
         self.layers.set_project(QgsProject.instance())
 
         self.lwc_versions[LwcVersions.Lizmap_3_4] = [